% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deworming_1_14y.R
\name{deworming_1_14y}
\alias{deworming_1_14y}
\title{Compute the number of deworming population target}
\format{
A numeric value
}
\usage{
deworming_1_14y(population, target_proportion = 0.493)
}
\arguments{
\item{population}{population estimate}

\item{target_proportion}{target proportion of the population estimate if different from default}
}
\description{
The number of deworming population target are computed as a proportion of the population estimate
}
\details{
At the time of developing this package the  deworming population target proportion was 49.30\% based on UNEPI standards
}
\examples{
\dontrun{
# Deworming population target by 2023

deworming_1_14y(45508000)
}
}
