% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girls_10y.R
\name{girls_10y}
\alias{girls_10y}
\title{Compute the number of Girls 10 years  population target}
\format{
A numeric value
}
\usage{
girls_10y(population, target_proportion = 0.0153)
}
\arguments{
\item{population}{population estimate}

\item{target_proportion}{target proportion of the population estimate if different from default}
}
\description{
The number of Girls 10 years  population target are computed as a proportion of the population estimate
}
\details{
At the time of developing this package the  Girls 10 years  population target proportion was 1.53\% based on UNEPI standards
for HPV Immunization
}
\examples{
\dontrun{
# Girls 10 years  population target by 2023

girls_10y(45508000)
}
}
