% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_pregnant_women.R
\name{non_pregnant_women}
\alias{non_pregnant_women}
\title{Compute the number of non pregnant women  population target}
\format{
A numeric value
}
\usage{
non_pregnant_women(population, target_proportion = 0.18)
}
\arguments{
\item{population}{population estimate}

\item{target_proportion}{target proportion of the population estimate if different from default}
}
\description{
The number of non pregnant women  population target are computed as a proportion of the population estimate
}
\details{
At the time of developing this package the  non pregnant women  population target proportion was 18\% based on UNEPI standards
for TT Immunization
}
\examples{
\dontrun{
# non pregnant women  population target by 2023

non_pregnant_women(45508000)
}
}
