% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pregnant_women.R
\name{pregnant_women}
\alias{pregnant_women}
\title{Compute the number of pregnant women  population target}
\format{
A numeric value
}
\usage{
pregnant_women(population, target_proportion = 0.05)
}
\arguments{
\item{population}{population estimate}

\item{target_proportion}{target proportion of the population estimate if different from default}
}
\description{
The number of pregnant women  population target are computed as a proportion of the population estimate
}
\details{
At the time of developing this package the  pregnant women  population target proportion was 5\% based on UNEPI standards
for TT Immunization
}
\examples{
\dontrun{
# pregnant women  population target by 2023

pregnant_women(45508000)
}
}
