% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surviving_infants_0_11m.R
\name{surviving_infants_0_11m}
\alias{surviving_infants_0_11m}
\title{Compute the number of surviving infants}
\format{
A numeric value
}
\usage{
surviving_infants_0_11m(population, target_proportion = 0.043)
}
\arguments{
\item{population}{population estimate}

\item{target_proportion}{target proportion of the population estimate if different from default}
}
\description{
The number of surviving infants are computed as a proportion of the population estimate
}
\details{
At the time of developing this package the surviving infants proportion was 4.30\% based on UNEPI standards
}
\examples{
\dontrun{
# surviving infants population target by 2023

surviving_infants_0_11m(45508000)
}
}
